SSR metadata in spot detail page
